user delete done

+++ backend/controllers/taskController.js
@@ -1,11 +1,12 @@
 const expressAsyncHandler = require("express-async-handler")
 const Task = require("../models/taskModel")
+const User = require("../models/userModel")
 
 //!@desc get task
 //!@route GET /api/task
 //!@access private
 const getTasks = expressAsyncHandler(async (req, res) => {
-  const task = await Task.find({ user: req.user.id })
+  const task = await Task.find({ user: req.user._id })
   res.status(200).json(task)
 })
 
@@ -18,11 +19,12 @@ const postTask = expressAsyncHandler(async (req, res) => {
     throw new Error("Add Task")
   }
 
-  const task = await Task.find({ task: req.body.task })
-  if (task) {
-    res.status(400)
-    throw new Error("Task Exists")
-  }
+  // const task = await Task.find({ task: req.body.task })
+
+  // if (task) {
+  //   res.status(400)
+  //   throw new Error(task, "Task Exists")
+  // }
 
   const newTask = await Task.create({ task: req.body.task, user: req.user._id })
   res.status(200).json(newTask)
@@ -32,17 +34,22 @@ const postTask = expressAsyncHandler(async (req, res) => {
 //!@route PUT /api/task/:id
 //!@access private
 const putTask = expressAsyncHandler(async (req, res) => {
-  const task = await Task.findone({ _id: req.params.id, user: req.user.id })
+  const task = await Task.findById(req.params.id)
+  const user = await User.findById(req.user.id)
+
   if (!task) {
     res.status(404)
     throw new Error("Task not found")
   }
 
-  const updatedTask = await Task.findByIdAndUpdate(
-    req.params.id,
-    { ...task, task: req.body.task },
-    { new: true }
-  )
+  if (user._id.toString !== task.user) {
+    res.status(401)
+    throw new Error("Unauthorized")
+  }
+
+  const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+  })
 
   res.status(200).json(updatedTask)
 })
@@ -51,15 +58,19 @@ const putTask = expressAsyncHandler(async (req, res) => {
 //!@route DELETE /api/task/:id
 //!@access private
 const deleteTask = expressAsyncHandler(async (req, res) => {
-  const task = await Task.findOne({
-    user: req.user.id,
-    _id: req.params.id,
-  })
+  const task = await Task.findById(req.params.id)
+  const user = await User.findById(req.user.id)
 
   if (!task) {
     res.status(404)
     throw new Error("Task not found")
   }
+
+  if (user._id.toString !== task.user) {
+    res.status(401)
+    throw new Error("Unauthorized")
+  }
+
   task.delete()
   res.status(200).json(task)
 })
